Return -1 for median val if nBlockHeight == 0
nBlockHeight gets updated only after a block gets processed in the fee
estimation object, in `BlockPolicyEstimator.ProcessBlock()`. This can happen
when the node has been started up new and clients are pounding on the
server node for fee estimations that it cannot currently
calculate.

The source of the block height comes from MempoolSignaled, which gets
block at boot time. So, as such, this is just a temporary fix when it
should be loading up the block correctly.