Add a new contributing document
While seeking contributors, it helps to set some rules to proceed with
development that we can all follow. The process is used to help get
everyone on the same page, but for now, without a lot of burden to
contribute.

I have opted to write a custom contributing document instead of
copy/pasting Bitcoin's. However, I used their document as inspiration,
since most of what is said, though specific in regards to Bitcoin,
applies very much to all large software projects.