getblocktemplate: txid and hash should be same endianness
This is the first step to fixing the problem completely. Other mining
pools tried to use the txid, but since it was in the wrong endianness,
it was hashing merkle trees wrong creating mismatch errors on merkle
roots when blocks were submitted.

This worked fine for miningcore, since it was only using the hash field.